+++ examples/sick_scan_rest_client_example.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023 SICK AG
+# Copyright (c) 2023 - 2024 SICK AG
 # SPDX-License-Identifier: MIT
 #
 
@@ -23,83 +23,85 @@ import json
 # - Write variable "ScanDataEnable".
 # - Read variable "ScanDataEnable".
 # - Call method "FindMe".
+# - Call the method "WriteEeprom".
+
 
 # Create the client object
-client = RESTClient(deviceIpAddress="192.168.0.1")
+client = RESTClient(device_ip_address="192.168.0.1")
 
 
 # Read the variable "ScanDataEthSettings".
 # The result is returned as a dictionary.
-variableName = "ScanDataEthSettings"
-success, resultScanDataEthSettings = client.readVariable(variableName)
-resultJSON = json.dumps(resultScanDataEthSettings) # The result is a dictionary which can be converted with json.dumps to a json string.
-print(f"Reading {variableName}: success = {success}, result={resultJSON}\n")
+variable_name = "ScanDataEthSettings"
+success, result_scan_data_eth_settings = client.read_variable(variable_name)
+result_json = json.dumps(result_scan_data_eth_settings) # The result is a dictionary which can be converted with json.dumps to a json string.
+print(f"Reading {variable_name}: success = {success}, result={result_json}\n")
 
 # Read the non existing variable "ScanDataEthSettingsTypo".
 # Success will be false and the corresponding error can be found in the result dictionary.
-variableName = "ScanDataEthSettingsTypo"
-success, result = client.readVariable(variableName)
-resultJSON = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
-print(f"Reading {variableName}: success = {success}, result={resultJSON}\n")
+variable_name = "ScanDataEthSettingsTypo"
+success, result = client.read_variable(variable_name)
+result_json = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
+print(f"Reading {variable_name}: success = {success}, result={result_json}\n")
 
 # Login with the user level "Service" and the corresponding password. Here we use the default password.
 # To login with other user levels please note that they have to be activated first via the UI.
-userLevel = "Service"
+userLevel = RESTClient.UserLevel.Service
 password = "servicelevel"
-success = client.setUserLevel(username=userLevel, password=password)
+success = client.set_user_level(userLevel, password)
 if success:
     print(f"Logging in with user level: {userLevel}\n")
 else:
-    print(f"Error logging in with user level: {userLevel}\n") 
+    print(f"Error logging in with user level: {userLevel}\n")
 
 # Write the variable "ScanDataEthSettings" and change ip address using a JSON string.
-variableName = "ScanDataEthSettings"
+variable_name = "ScanDataEthSettings"
 body = '{"Protocol": 1, "IPAddress": [192, 168, 0, 100], "Port": 2115}'
-success, result = client.writeVariable(variableName, json.loads(body))
-resultJSON = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
-print(f"Writing {variableName} based on JSON string as body: success = {success}, result={resultJSON}\n")
+success, result = client.write_variable(variable_name, json.loads(body))
+result_json = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
+print(f"Writing {variable_name} based on JSON string as body: success = {success}, result={result_json}\n")
 
 # Write the variable "ScanDataEthSettings" and change ip address using the previously read dictionary.
-variableName = "ScanDataEthSettings"
-body = resultScanDataEthSettings["data"][variableName]
+variable_name = "ScanDataEthSettings"
+body = result_scan_data_eth_settings["data"][variable_name]
 body["IPAddress"] = [192,168,0,102]
-success, result = client.writeVariable(variableName, body)
-resultJSON = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
-print(f"Writing {variableName} using previously read body: success = {success}, result={resultJSON}\n")
+success, result = client.write_variable(variable_name, body)
+result_json = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
+print(f"Writing {variable_name} using previously read body: success = {success}, result={result_json}\n")
 
 # Read the variable "ScanDataEthSettings" and make sure that value changed.
-variableName = "ScanDataEthSettings"
-success, result = client.readVariable(variableName)
-resultJSON = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
-print(f"Reading {variableName} with changed ip address: success = {success}, result={resultJSON}\n")
+variable_name = "ScanDataEthSettings"
+success, result = client.read_variable(variable_name)
+result_json = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
+print(f"Reading {variable_name} with changed ip address: success = {success}, result={result_json}\n")
 
 # Write the variable "ScanDataEnable".
 # This is an example how to write a variable with a single value as argument.
-variableName = "ScanDataEnable"
+variable_name = "ScanDataEnable"
 body = True
-success, result = client.writeVariable(variableName, body)
-resultJSON = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
-print(f"Writing {variableName} with parameter {body}: success = {success}, result={resultJSON}\n")
+success, result = client.write_variable(variable_name, body)
+result_json = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
+print(f"Writing {variable_name} with parameter {body}: success = {success}, result={result_json}\n")
 
 # Read variable "ScanDataEnable".
-variableName = "ScanDataEnable"
-success, result = client.readVariable(variableName)
-resultJSON = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
-print(f"Reading {variableName}: success = {success}, result={resultJSON}\n")
+variable_name = "ScanDataEnable"
+success, result = client.read_variable(variable_name)
+result_json = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
+print(f"Reading {variable_name}: success = {success}, result={result_json}\n")
 
 # Call the method "FindMe". Both device LEDs will blink for the given duration (in seconds).
-methodName = "FindMe"
+method_name = "FindMe"
 body =  {"uiDuration": 5}
-success, result = client.callMethod(methodName, body)
-resultJSON = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
-print(f"Calling {methodName} with parameter {body}: success = {success}, result={resultJSON}\n")
+success, result = client.call_method(method_name, body)
+result_json = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
+print(f"Calling {method_name} with parameter {body}: success = {success}, result={result_json}\n")
 
 # Call the method "WriteEeprom".
 # This is to show how to call a method without a parameter.
 # Please note that calling this method saves the previously written device parameters permanently!
 # Above the data output parameters are modified. Change them back and save parameters again if required.
-methodName = "WriteEeprom"
+method_name = "WriteEeprom"
 body =  None
-success, result = client.callMethod(methodName, body)
-resultJSON = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
-print(f"Calling {methodName} with no parameter: success = {success}, result={resultJSON}\n")
+success, result = client.call_method(method_name, body)
+result_json = json.dumps(result) # The result is a dictionary which can be converted with json.dumps to a json string.
+print(f"Calling {method_name} with no parameter: success = {success}, result={result_json}\n")